// src/data/api/generated/models/guestSavedLocation.ts
/* eslint-disable */
// @ts-nocheck

export interface GuestSavedLocation {
  /** @nullable */
  address_line?: string | null;
  branch_id: string;
  channel: string;
  /** @nullable */
  customer_lat?: number | null;
  /** @nullable */
  customer_lng?: number | null;
  /** @nullable */
  floor?: string | null;
  /** @nullable */
  landmark?: string | null;
  last_used_at: string;
  /** @nullable */
  place_name?: string | null;
  /** @nullable */
  unit_number?: string | null;
}